review

+++ cent_core/center_loss.py
@@ -14,6 +14,6 @@ class CenterLoss(torch.nn.Module):
         xs = torch.nn.functional.normalize(xs)
         center_exp = self.center.index_select(dim=0, index=ys.long())
         count = torch.histc(ys.float(), bins=self.cls_num, min=0, max=self.cls_num - 1)
-        count_exp = count.index_select(dim=0, index=ys.long())
+        count_exp = count.index_select(dim=0, index=ys.long()).float()
         loss = torch.mean(torch.sum(torch.pow(xs - center_exp, 2) / 2, dim=1) / count_exp)
         return loss
